agent/requirements.txt
botocore
bedrock-agentcore

strands-agents
strands-agents[mcp]
strands-agents[otel]
